% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_verifier_simple_server.R
\name{app_verifier_simple_server}
\alias{app_verifier_simple_server}
\title{main server function}
\usage{
app_verifier_simple_server(input, output, session)
}
\description{
main server function
}
