% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_controls_simpleUI.R
\name{create_controls_simpleUI}
\alias{create_controls_simpleUI}
\title{A module to add controls like nav buttons for the spectrogram part of the app}
\usage{
create_controls_simpleUI(id)
}
\arguments{
\item{id, }{id}
}
\description{
A module to add controls like nav buttons for the spectrogram part of the app
}
