% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_controlsUI.R
\name{create_controlsUI}
\alias{create_controlsUI}
\title{A module to add controls like nav buttons for the spectrogram part of the app}
\usage{
create_controlsUI(id)
}
\arguments{
\item{id, }{id}
}
\description{
A module to add controls like nav buttons for the spectrogram part of the app
}
