% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_create.R
\name{database_create}
\alias{database_create}
\title{create an empty SQLite database with a table to hold verification decisions}
\usage{
database_create(path_db, name_db = "verifications.sqlite")
}
\description{
create an empty SQLite database with a table to hold verification decisions
}
