% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_export.R
\name{database_export}
\alias{database_export}
\title{export database}
\usage{
database_export(con, file_db)
}
\arguments{
\item{con}{= database connection object}

\item{file_db}{= reactive object giving path and name of database}
}
\description{
export database
}
