% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_button_toggles.R
\name{nav_button_toggles}
\alias{nav_button_toggles}
\title{Toggle enable/disable of nav buttons depending on state}
\usage{
nav_button_toggles(numfiles, nthfile)
}
\arguments{
\item{numfiles}{= the total number of audio files}

\item{nthfile}{= which file is currently active}
}
\description{
Toggle enable/disable feature of navigation buttons
}
