% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_full_clip_verifier.R
\name{run_full_clip_verifier}
\alias{run_full_clip_verifier}
\title{run the full verification app}
\usage{
run_full_clip_verifier(audiodirs = NULL)
}
\arguments{
\item{audiodirs}{= named list providing alternative drives and shortcuts to audio directories}
}
\description{
run the full verification app
}
