% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simple_clip_verifier.R
\name{run_simple_clip_verifier}
\alias{run_simple_clip_verifier}
\title{run the simple clip verifier app}
\usage{
run_simple_clip_verifier(audiodirs = NULL, choices = NULL)
}
\arguments{
\item{audiodirs}{= named list providing alternative drives and shortcuts}

\item{choices}{= optional list to populate verification buttons. Defaults to 
True, False and Unknown if not provided}
}
\description{
run the simple clip verifier app
}
