% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_fast.R
\name{spec_fast}
\alias{spec_fast}
\title{Produce a spectrogram quickly!}
\usage{
spec_fast(signal, sr, window_size, overlap, theme, ylim, xmin, xmax)
}
\arguments{
\item{window_size}{= window size for the FFT}

\item{overlap}{= proportional overlap of successive windows}

\item{theme}{= Greyscale or Viridis}

\item{ylim}{= vector for min and max of the y-axis range}

\item{path_to_wav}{= full path to the audio file}
}
\description{
Reads an audio file and produces a spectrogram based on given window 
length and overlap. Most of the function is written from first principles and 
librosa methods from Python. Some inspiration for the image plotting from 
Aaron Albin's Spectrogram() function.
}
