% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_fast4labeller.R
\name{spec_fast4labeller}
\alias{spec_fast4labeller}
\title{Produce a spectrogram quickly!}
\usage{
spec_fast4labeller(
  signal,
  sr,
  window_size,
  overlap,
  theme,
  ylim,
  xmin = 0,
  xmax
)
}
\arguments{
\item{window_size}{= window size for the FFT}

\item{overlap}{= proportional overlap of successive windows}

\item{theme}{= Greyscale or Viridis}

\item{ylim}{= vector for min and max of the y-axis range}

\item{path_to_wav}{= full path to the audio file}
}
\description{
Reads an audio file and produces a spectrogram based on given window
length and overlap. Most of the function is written from first principles and
librosa methods from Python. Some inspiration for the image plotting from
Aaron Albin's Spectrogram() function.
}
