% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_decision.R
\name{submit_decision}
\alias{submit_decision}
\title{Function to submit the verification decision}
\usage{
submit_decision(
  state,
  input,
  session,
  splist,
  batch_params,
  path_audio,
  con,
  recent_labels,
  for_training = FALSE
)
}
\arguments{
\item{for_training, }{bool, whether the clip should be nominated for training}
}
\description{
Function to submit the verification decision
}
