% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_title_panel.R
\name{ui_title_panel}
\alias{ui_title_panel}
\title{Add BTO title panel}
\usage{
ui_title_panel(app_title)
}
\arguments{
\item{app_title}{= str, the name of the app to be shown in the title panel}
}
\description{
Add BTO title panel
}
